+++ forge/src/main/java/net/mehvahdjukaar/moonlight/core/mixins/forge/FireBlockMixin.java
@@ -1,5 +1,7 @@
 package net.mehvahdjukaar.moonlight.core.mixins.forge;
 
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
+import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
 import com.llamalad7.mixinextras.sugar.Local;
 import com.llamalad7.mixinextras.sugar.Share;
 import com.llamalad7.mixinextras.sugar.ref.LocalRef;
@@ -25,15 +27,16 @@ public abstract class FireBlockMixin extends BaseFireBlock {
         super(settings, damage);
     }
 
-    @Inject(method = "tryCatchFire",
+    @WrapOperation(method = "tryCatchFire",
             at = @At(value = "INVOKE",
-                    target = "net/minecraft/world/level/Level.removeBlock (Lnet/minecraft/core/BlockPos;Z)Z",
-                    shift = At.Shift.AFTER))
-    private void afterRemoveBlock(Level level, BlockPos pos, int chance, RandomSource pRandom, int age,
-                                  Direction face, CallbackInfo ci, @Local BlockState before) {
-        var event = IFireConsumeBlockEvent.create(pos, level, before, chance, age, face);
+                    target = "net/minecraft/world/level/Level.removeBlock (Lnet/minecraft/core/BlockPos;Z)Z"))
+    private boolean afterRemoveBlock(Level level, BlockPos pos, boolean isMoving, Operation<Boolean> original,
+                                     @Local BlockState oldState, @Local(ordinal = 0, argsOnly = true) int chance,
+                                     @Local(ordinal = 1, argsOnly = true) int age, @Local(argsOnly = true) Direction face) {
+        var event = IFireConsumeBlockEvent.create(pos, level, oldState, chance, age, face);
         MoonlightEventsHelper.postEvent(event, IFireConsumeBlockEvent.class);
         BlockState newState = event.getFinalState();
-        if (newState != null) level.setBlockAndUpdate(pos, newState);
+        if (newState != null) return level.setBlockAndUpdate(pos, newState);
+        else return original.call(level, pos, isMoving);
     }
 }